pyproject.toml
[tool.poetry]
name = "eflips-book"
version = "0.1.0"
description = "Sample Scripts for the eFLIPS book"
authors = ["Ludger Heide <ludger.heide@tu-berlin.de>"]
license = "AGPLv3"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
eflips-depot = "^4.0.0"
eflips-ingest = "^1.2.4"
eflips-model = "^4.0.0a1"
eflips-eval = "^1.3.0"
jupyter = "^1.0.0"
matplotlib = "^3.9.1"
seaborn = "^0.13.2"
tqdm = "^4.66.4"
eflips-opt = {git = "git@github.com:mpm-tu-berlin/eflips-opt.git", rev = "bugfix/eflips-model-version"}
kaleido = "0.2.1"
django-simba = { git = "git@github.com:rl-institut/django-simba.git", branch = "feature/djago-simba-as-package" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
